requirements.txt
APScheduler==3.9.1.post1
asgiref==3.6.0
beautifulsoup4==4.11.1
branca==0.6.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.2.5
django-apscheduler==0.6.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
folium==0.14.0
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
numpy==1.25.2
openpyxl==3.1.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
